+++ src/pages/google-oauth-callback/index.ts
@@ -1,1 +1,0 @@
-export { parseGoogleOAuthCallbackParams } from './model/parseGoogleOAuthCallbackParams'